requirements.txt
djangorestframework==3.14.0
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.2

